+++ package.json
@@ -10,6 +10,10 @@
     "lint": "eslint src/",
     "docd": "docker-compose -f docker-compose.yml -f docker-compose.dev.yml"
   },
+  "private": true,
+  "workspaces": [
+    "packages/*"
+  ],
   "husky": {
     "hooks": {
       "pre-commit": "lint-staged"
@@ -23,18 +27,7 @@
   },
   "keywords": [],
   "dependencies": {
-    "@okgrow/graphql-scalars": "^0.4.2",
-    "axios": "^0.18.0",
-    "bull": "^3.4.8",
-    "dataloader": "^1.4.0",
-    "fs-extra": "^7.0.0",
-    "glob": "^7.1.3",
-    "graphql-yoga": "1.16.2",
-    "ioredis": "^4.2.0",
-    "lodash": "^4.17.11",
-    "merge-graphql-schemas": "^1.5.7",
-    "path": "^0.12.7",
-    "pino": "^5.8.0"
+    "@openapi/core": "1.0.0"
   },
   "devDependencies": {
     "babel-eslint": "^10.0.1",
