+++ packages/common/common.resolver.js
@@ -1,0 +1,7 @@
+// const providers = require('../');
+
+module.exports = {
+  Query: {
+    providers: (obj, args, context, info) => [{ name: 'hop' }],
+  },
+};
