+++ packages/openlibrary/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@openapi/plugin-openlibrary",
+  "version": "1.0.0",
+  "license": "MIT",
+  "dependencies": {
+    "@openapi/core": "1.0.0",
+    "axios": "^0.18.0",
+    "dataloader": "^1.4.0",
+    "lodash": "^4.17.11"
+  }
+}
