+++ packages/openweather/constants.js
