+++ packages/openweather/openweather.type.graphql
