+++ packages/openweather/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@openapi/plugin-openweather",
+  "version": "1.0.0",
+  "license": "MIT",
+  "dependencies": {
+    "@openapi/core": "1.0.0",
+    "axios": "^0.18.0"
+  }
+}
