--- src/providers/common/common.resolver.js
@@ -1,7 +1,0 @@
-const providers = require('../');
-
-module.exports = {
-  Query: {
-    providers: (obj, args, context, info) => providers,
-  },
-};
